Gen/param entity basic parsing
of course have not achieve external parsing yet. Entities are going to be a challenge to integrate into the parser but it needs to be done to form a somewhat conforming validating parser. Hopefully can get entities fully integrated within a few days. And then semantic validation may be one of the final stages before the parser is nearly done.